# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Bilingual News Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sources:
    get:
      operationId: listSources
      tags: [sources]
      summary: List news sources
      responses:
        "200":
          description: Sources list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsSource"
    post:
      operationId: createSource
      tags: [sources]
      summary: Add news source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsSourceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsSource"
  /sources/{id}:
    patch:
      operationId: updateSource
      tags: [sources]
      summary: Update source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsSourceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsSource"
    delete:
      operationId: deleteSource
      tags: [sources]
      summary: Delete source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /sources/{id}/fetch:
    post:
      operationId: fetchFromSource
      tags: [sources]
      summary: Manually trigger fetch from a source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fetch result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FetchResult"
...
components:
  schemas:
...
    NewsSource:
      type: object
      required: [id, name, url, lang, active, articleCount]
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        feedUrl:
          type: ["string", "null"]
        lang:
          type: string
          enum: [ar, en]
        active:
          type: boolean
        categoryId:
          type: ["integer", "null"]
        articleCount:
          type: integer
        lastFetchedAt:
          type: ["string", "null"]
        fetchInterval:
          type: integer
        reliability:
          type: integer
    NewsSourceInput:
      type: object
      required: [name, url, lang]
      properties:
        name:
          type: string
        url:
          type: string
        feedUrl:
          type: ["string", "null"]
        lang:
          type: string
          enum: [ar, en]
        active:
          type: boolean
        categoryId:
          type: ["integer", "null"]
        fetchInterval:
          type: integer
    NewsSourceUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        feedUrl:
          type: ["string", "null"]
        lang:
          type: ["string", "null"]
        active:
          type: ["boolean", "null"]
        categoryId:
          type: ["integer", "null"]
        fetchInterval:
          type: ["integer", "null"]
...
    FetchResult:
      type: object
      required: [success, articlesAdded]
      properties:
        success:
          type: boolean
        articlesAdded:
          type: integer
        message:
          type: ["string", "null"]