# lib/api-zod/src/generated/types/fetchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Bilingual News Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface FetchResult {
  success: boolean;
  articlesAdded: number;
  /** @nullable */
  message?: string | null;
}